moved Link component to Extensions
