+++ src/components/common/ColonyHome/ColonyUnclaimedTransfers/ColonyUnclaimedTransfers.tsx
@@ -6,7 +6,7 @@ import Heading from '~shared/Heading';
 import NavLink from '~shared/NavLink';
 import Numeral from '~shared/Numeral';
 import { Tooltip } from '~shared/Popover';
-import Link from '~shared/Link';
+import Link from '~shared/Extensions/Link';
 
 import { ActionTypes } from '~redux';
 import { useColonyContext, useColonyFundsClaims } from '~hooks';
