+++ src/components/common/Extensions/CalamityBanner/CalamityBanner.tsx
@@ -1,7 +1,7 @@
 import React, { FC, PropsWithChildren } from 'react';
 import { useIntl } from 'react-intl';
 import { CalamityBannerProps } from './types';
-import Link from '~shared/Link';
+import Link from '~shared/Extensions/Link';
 import Button from '~shared/Extensions/Button';
 import Icon from '~shared/Icon';
 import { useCalamityBanner } from './hooks';
