+++ src/components/common/Extensions/UserNavigation/partials/UserMenu/UserMenu.tsx
@@ -4,13 +4,13 @@ import { useIntl } from 'react-intl';
 import { UserMenuProps } from './types';
 import { useAppContext, useMobile } from '~hooks';
 import Button from '~shared/Extensions/Button';
-import Link from '~shared/Link';
 import Icon from '~shared/Icon';
 import ThemeSwitcher from '~common/Extensions/ThemeSwitcher';
 import styles from './UserMenu.module.css';
 import PopoverBase from '~shared/Extensions/PopoverBase';
 import TitledContent from '~common/Extensions/TitledContent/TitledContent';
 import WalletConnectedTopMenu from '../WalletConnectedTopMenu/WalletConnectedTopMenu';
+import Link from '~shared/Extensions/Link';
 
 const displayName = 'common.Extensions.UserNavigation.partials.UserMenu';
