+++ src/components/common/Extensions/UserNavigation/partials/WalletPopover/WalletPopover.tsx
@@ -3,7 +3,7 @@ import { useIntl } from 'react-intl';
 import { WalletPopoverProps } from './types';
 import PopoverBase from '~shared/Extensions/PopoverBase';
 import Icon from '~shared/Icon';
-import Link from '~shared/Link';
+import Link from '~shared/Extensions/Link';
 import styles from './WalletPopover.module.css';
 import WalletPopoverOption from '../WalletPopoverOption';
 import Button from '~shared/Extensions/Button';
