+++ src/components/shared/Extensions/Toast/Toast.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import clsx from 'clsx';
-import Link from '~shared/Link';
+import Link from '~shared/Extensions/Link';
 import { ToastProps } from './types';
 import Icon from '~shared/Icon';
